added admin check and new help messages

+++ src/handlers/admins.ts
@@ -1,4 +1,5 @@
 import { bot } from "index";
+import { lib } from "lib/lib";
 import { Message } from "node-telegram-bot-api";
 import { adminsServices } from "services/Admins";
 
@@ -14,6 +15,7 @@ export async function createDefaultAdmins() {
 export const adminsHandlers = {
     async createAdmin(message: Message): Promise<void> {
         try {
+            if(!lib.isAdmin(message)) return;
             await adminsServices.createAdmin(message.chat.id);
             bot.sendMessage(message.chat.id, "Администратор создан успешно");
         } catch (e) {
@@ -22,6 +24,7 @@ export const adminsHandlers = {
     },
     async deleteAdmin(message: Message): Promise<void> {
         try {
+            if(!lib.isAdmin(message)) return;
             await adminsServices.deleteAdmin(message.chat.id);
             bot.sendMessage(message.chat.id, "Администратор удалён успешно");
         } catch (e) {
